requirements.txt
fastapi
uvicorn
jinja2
python-dotenv
pymysql
boto3
